trusted-listener/requirements.txt
PyYAML==6.0.1
Rtree==1.2.0
watchdog==4.0.0
